[FEATURE] Added logic for scale

+++ src/handlers.js
@@ -5,7 +5,6 @@ const AWS = require('aws-sdk');
 const sharp = require('sharp');
 const { v4: uuidv4 } = require('uuid');
 const { run: runSiteScroll } = require('./site_scroll.js');
-const path = require('path');
 
 const { AWS_KEY_ID, AWS_SECRET_KEY, AWS_S3_AD_VIDEOS_BUCKET } = process.env;
 
@@ -22,7 +21,8 @@ async function handle(templateName, data) {
     userId,
     callBackUrl,
     errorCallBackUrl,
-    productUrl
+    productUrl,
+    scaleAfter,
   } = data;
   console.time('BUILD_TIME');
   
@@ -46,9 +46,13 @@ async function handle(templateName, data) {
     const coverPath = await createCover(`${folderPath}/output.mp4`, folderPath);
     
     const fileName = uuidv4();
-    
+    if(scaleAfter) {
+      await runCommand(`ffmpeg -i ${folderPath}/output.mp4 -vf "scale=2160:3840" ${folderPath}/output_scale_command.mp4`);
+      fs.unlinkSync(`${folderPath}/output.mp4`);
+      fs.renameSync(`${folderPath}/output_scale_command.mp4`, `${folderPath}/output.mp4`);
+    }
+    await runCommand(`ffmpeg -i ${folderPath}/output.mp4 -vf "scale=2160:3840" ${folderPath}/output_scale_command.mp4`);
     console.timeEnd('BUILD_TIME');
-    
     const url = await uploadToS3(`${folderPath}/output.mp4`, `${userId}/${adVideoId}/${fileName}.mp4`);
     const coverUrl = await uploadToS3(coverPath, `${userId}/${adVideoId}/${fileName}_cover.png`, 'image/png');
 
@@ -66,12 +70,12 @@ async function handle(templateName, data) {
       status: 'error'
     })
   } finally {
-    fs.rm(folderPath, { recursive: true }, (err) => {
-      if (err) {
-        console.error(err);
-        return;
-      }
-    });
+    // fs.rm(folderPath, { recursive: true }, (err) => {
+    //   if (err) {
+    //     console.error(err);
+    //     return;
+    //   }
+    // });
   }
 }
 
@@ -80,7 +84,10 @@ async function prepare(folderPath, data) {
     media,
     avatarUrl,
     actionUrl,
-    musicUrl
+    musicUrl,
+    width = 2160,
+    height = 3840,
+    scaleAfter = false
   } = data;
   
   const videos = media.filter(mediaItem => mediaItem.type === 'video');
@@ -100,7 +107,7 @@ async function prepare(folderPath, data) {
       await download(mediaItem.url, mediaPathOrg);
   
       await sharp(mediaPathOrg)
-      .resize(4320)
+      .resize(width * 2)
       .jpeg({ quality: 100 })
       .toFile(mediaPath);
       fs.unlinkSync(mediaPathOrg);
@@ -110,12 +117,14 @@ async function prepare(folderPath, data) {
 
   await runCommand(`ffmpeg -i ${folderPath}/avatar.mp4  -vf "setpts=${1/acceleration}*PTS" -filter:a "atempo=${acceleration}" -q:v 3 -q:a 3 ${folderPath}/avatar_speed.mp4`);
 
-  await runCommand(`ffmpeg -i ${folderPath}/avatar_speed.mp4 -filter:a "volume=2.0" ${folderPath}/avatar_speed_sound.mp4`);
-
-  await runCommand(`ffmpeg -i ${folderPath}/avatar_speed_sound.mp4 -vf "scale=2160:3840" ${folderPath}/avatar_end.mp4`);
-  
+  if(!scaleAfter) {
+    await runCommand(`ffmpeg -i ${folderPath}/avatar_speed.mp4 -filter:a "volume=2.0" ${folderPath}/avatar_speed_sound.mp4`);
+    await runCommand(`ffmpeg -i ${folderPath}/avatar_speed_sound.mp4 -vf "scale=${width}:${height}" ${folderPath}/avatar_end.mp4`);
+    fs.unlinkSync(`${folderPath}/avatar_speed_sound.mp4`);
+  } else {
+    await runCommand(`ffmpeg -i ${folderPath}/avatar_speed.mp4 -filter:a "volume=2.0" ${folderPath}/avatar_end.mp4`);
+  }
   fs.unlinkSync(`${folderPath}/avatar_speed.mp4`);
-  fs.unlinkSync(`${folderPath}/avatar_speed_sound.mp4`);
   fs.renameSync(`${folderPath}/avatar.mp4`, `${folderPath}/avatar_org.mp4`);
   fs.renameSync(`${folderPath}/avatar_end.mp4`, `${folderPath}/avatar.mp4`);
 
@@ -147,110 +156,6 @@ async function generateSubtitles(folderPath) {
     }
   });
   return `${folderPath}/transcription.json`;
-}
-
-async function generateSubtitlesV2(folderPath, subtitlesUrl) {
-  await download(subtitlesUrl, `${folderPath}/subtitles.srt`);
-  const inputFilePath = path.join(folderPath, 'subtitles.srt');
-  const jsonArray = readFileAndParse(inputFilePath, folderPath);
-  console.error(jsonArray)
-  const words = calculateWordTimings(jsonArray);
-  fs.writeFileSync(`${folderPath}/transcription.json`, JSON.stringify(words), { encoding: 'utf8' });
-}
-
-function calculateWordTimings(subtitles) {
-  const wordTimings = [];
-  console.log(subtitles)
-  subtitles.forEach(subtitle => {
-    const words = subtitle.text.split(' ');
-    const totalDuration = subtitle.end - subtitle.start;
-    const durationPerLetter = totalDuration / subtitle.text.length;
-
-    let currentStart = subtitle.start;
-
-    words.forEach(word => {
-      const wordDuration = word.length * durationPerLetter;
-      const wordEnd = currentStart + wordDuration;
-
-      wordTimings.push({
-        start: currentStart,
-        end: wordEnd,
-        word: word
-      });
-
-      currentStart = wordEnd;
-    });
-  });
-
-  return wordTimings;
-}
-
-function readFileAndParse(filePath, folderPath) {
-    const data = fs.readFileSync(filePath)
-
-    const jsonArray = parseSRT(data);
-
-    return jsonArray;
-}
-
-function parseSRT(data) {
-  const srtRegex = /(\d+)\n(\d{2}:\d{2}:\d{2},\d{3}) --> (\d{2}:\d{2}:\d{2},\d{3})\n([\s\S]*?)(?=\n\n|\n*$)/g;
-  const result = [];
-  let match;
-
-  while ((match = srtRegex.exec(data)) !== null) {
-    result.push({
-      start: convertTimeToSeconds(match[2]),
-      end: convertTimeToSeconds(match[3]),
-      text: match[4].replace(/\n/g, ' ')
-    });
-  }
-
-  return result;
-}
-
-function convertTimeToSeconds(time) {
-  const [hours, minutes, seconds] = time.split(':');
-  const [secs, millis] = seconds.split(',');
-  return parseInt(hours, 10) * 3600 + parseInt(minutes, 10) * 60 + parseInt(secs, 10) + parseInt(millis, 10) / 1000;
-}
-
-async function getCorrectedSubtitles(originalText, subtitles) {
-  const prompt = `
-  Given the original text and the subtitles JSON, correct the subtitles to match the original text perfectly. Ensure there are no errors and return only the corrected subtitles in JSON format.
-
-  Original Text:
-  ${originalText}
-
-  Subtitles:
-  ${JSON.stringify(subtitles)}
-
-  Corrected Subtitles (JSON only):
-  `;
-
-  try {
-      const response = await axios.post('https://api.openai.com/v1/chat/completions', {
-          model: 'gpt-4',
-          messages: [
-              { role: 'system', content: 'You are a helpful assistant.' },
-              { role: 'user', content: prompt }
-          ],
-          max_tokens: 4096
-      }, {
-          headers: {
-              'Authorization': `Bearer ${OPENAI_API_KEY}`,
-              'Content-Type': 'application/json'
-          }
-      });
-
-      const correctedSubtitles = response.data.choices[0].message.content;
-      console.error('Corrected Subtitles:', JSON.stringify(response.data));
-      return correctedSubtitles;
-
-  } catch (error) {
-      console.error('Error while calling ChatGPT API:', error);
-      throw error;
-  }
 }
 
 async function createCover(videoPath, folderPath) {
